stock-auth

+++ app/Http/Controllers/ProductController.php
@@ -8,27 +8,44 @@ use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Support\Facades\Storage;
 use Image;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
+//construc for curruent user
+
+public $user;
+public function __construct()
+{
+    $this->middleware(function($request,$next){
+
+$this->user=Auth::user();
+return $next($request);
+    });
+}
+
+
+
   // add Product view page
   public function AddProduct()
   {
-    $categories = Category::latest()->get();  
+
+    $categories = Category::latest()->get();
       return view('product.AddProduct',compact('categories'));
   }
- 
- 
+
+
   // public function ProductList()
   // {
   //   $products=Product::all();
   //     return view('product.ProductList',compact('products'));
   // }
 
-  
+
   public function StoreProduct(Request $request)
   {
-    
+
+
        $validateData = $request->validate([
            'name' => 'required',
            'price' => 'required',
@@ -38,7 +55,7 @@ class ProductController extends Controller
 
 
 
-          
+
        ]);
       $image = $request->file('image');
       $name_gen = hexdec(uniqid()).'.'.$image->getClientOriginalExtension();
@@ -51,32 +68,40 @@ class ProductController extends Controller
       $product->product_code=$request->product_code;
       $product->squ_code=$request-> squ_code;
       $product->count=$request->count;
-     
-      $product->product_image= $save_url; 
-      $product->product_satus= 1; 
+
+      $product->product_image= $save_url;
+      $product->product_satus= 1;
       $product->save();
       $notification = array(
         'message' => 'Product Add Sucessyfuly',
         'alert-type' => 'success',
       );
-  
+
       return redirect()->back()->with($notification);
-  
+
     }
   public function showProduct(){
+
+
     $products = Product::all();
 
     return view('Product.ProductList', compact('products'));
 }
 public function EditProduct($id)
-{  
+{
+    if(is_null($this->user) || !$this->user->can('product.create') || !$this->user->can('product.update') || !$this->user->can('product.delete') || !$this->user->can('product.view')){
+        abort('403','You dont have acces!!!!');
+    }
     $product = Product::find($id);
-    $categories = Category::latest()->get(); 
+    $categories = Category::latest()->get();
     return view('product.ProductEdit',compact('product','categories'));
 }
 
 public function UpdateProduct(Request $request,$id)
-{    
+{
+    if(is_null($this->user) || !$this->user->can('product.create') || !$this->user->can('product.update') || !$this->user->can('product.delete') || !$this->user->can('product.view')){
+        abort('403','You dont have acces!!!!');
+    }
       $validateData = $request->validate([
     'name' => 'required',
     'price' => 'required',
@@ -86,7 +111,7 @@ public function UpdateProduct(Request $request,$id)
 
 
 
-   
+
 ]);
       $image = $request->file('image');
       $name_gen = hexdec(uniqid()).'.'.$image->getClientOriginalExtension();
@@ -99,20 +124,23 @@ public function UpdateProduct(Request $request,$id)
           $product->product_code=$request->product_code;
           $product->squ_code=$request-> squ_code;
           $product->count=$request->count;
-         
-          $product->product_image= $save_url; 
-          $product->product_satus= 1; 
+
+          $product->product_image= $save_url;
+          $product->product_satus= 1;
           $product->save();
           $notification = array(
             'message' => 'Product Edited Sucessyfuly',
             'alert-type' => 'success',
           );
-      
+
           return redirect()->back()->with($notification);
-      
+
         }
 public function DeleteProduct($id)
 {
+    if(is_null($this->user) || !$this->user->can('product.create') || !$this->user->can('product.update') || !$this->user->can('product.delete') || !$this->user->can('product.view')){
+        abort('403','You dont have acces!!!!');
+    }
     Product::destroy($id);
     $notification = array(
       'message' => 'Product deleted Sucessyfuly',
@@ -122,7 +150,7 @@ public function DeleteProduct($id)
     return redirect()->back()->with($notification);
 
   }
-    
+
 }
